[#139] fix: 파일 분리
